use hash
